Create add functionality

+++ src/Item.java
@@ -1,9 +1,9 @@
 public class Item {
     private int id;
     private String name;
-    private float price;
+    private double price;
 
-    public Item(int id, String name, float price) {
+    public Item(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -25,11 +25,11 @@ public class Item {
         this.name = name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
